+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add `Dune.string_to_quoted/2` to make it possible to visualize AST
 - Merged parsing and eval options in a single `Dune.Opts` for simplicity
+- Add a `pretty` option to inspect result
 - Better error message when `def/2` and `defp/2` called outside a module
 
 ### Breaking changes